+++ src/swap.h
@@ -5,7 +5,7 @@ static inline uint16_t ltohs(uint16_t x)
     /* This is compile-time optimised with at least -O1 or -Os */
     uint32_t const dummy = 0x12345678;
     if(*(uint8_t const *)&dummy == 0x12)
-        return (x >> 8) | (x << 8);
+        return (uint16_t) ((x >> 8) | (x << 8));
     else
         return x;
 }
